Extract `ClassHasStaticAttribute` tests into separate test case
